test(E2E): Move support files to support folder
