+++ test/e2e/cli-arguments.spec.js
@@ -9,7 +9,7 @@ Unless required by applicable law or agreed to in writing, software distributed 
 */
 
 const path = require("path");
-const { CliRunner, request, wait, TimeCounter, BINARY_PATH } = require("./utils");
+const { CliRunner, request, wait, TimeCounter, BINARY_PATH } = require("./support/utils");
 
 describe("command line arguments", () => {
   const cwdPath = path.resolve(__dirname, "fixtures");
