+++ test/e2e/programmatic-server.spec.js
@@ -8,7 +8,7 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-const { CliRunner, request, wait, TimeCounter, fixturesFolder } = require("./utils");
+const { CliRunner, request, wait, TimeCounter, fixturesFolder } = require("./support/utils");
 
 describe("programmatic Server", () => {
   const cwdPath = fixturesFolder("programmatic-server");
