+++ test/e2e/support/InteractiveCliRunner.js
@@ -11,7 +11,6 @@ Unless required by applicable law or agreed to in writing, software distributed 
 const { CliRunner, wait } = require("./utils");
 
 const LOG = "[CLI]: ";
-const SCREEN_SEPARATOR = ">> Mocks server";
 
 module.exports = class InteractiveCliRunner {
   constructor(cliArguments, cliOptions) {
@@ -24,11 +23,7 @@ module.exports = class InteractiveCliRunner {
 
   async getCurrentScreen() {
     await wait(500);
-    const allLogs = this._cli.allLogs;
-    const lastLog = allLogs[allLogs.length - 1];
-    return lastLog.includes(SCREEN_SEPARATOR)
-      ? `${SCREEN_SEPARATOR}${lastLog.split(SCREEN_SEPARATOR).pop()}`
-      : lastLog;
+    return this._cli.currentScreen;
   }
 
   async logCurrentScreen() {
@@ -57,6 +52,10 @@ module.exports = class InteractiveCliRunner {
 
   get logs() {
     return this._cli.logs;
+  }
+
+  get currentScreen() {
+    return this._cli.currentScreen;
   }
 
   async cursorDown(number) {
