+++ test/e2e/web-tutorial-json-admin-api.spec.js
@@ -8,7 +8,13 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-const { startServer, stopServer, request, changeBehavior, getBehaviors } = require("./utils");
+const {
+  startServer,
+  stopServer,
+  request,
+  changeBehavior,
+  getBehaviors,
+} = require("./support/utils");
 
 describe("web tutorial", () => {
   let server;
